# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/wa-base-times": {
      "get": {
        "operationId": "WaBaseTimesController_findAll",
        "parameters": [
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "poolLength",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List WA base times, optionally filtered by year and pool length",
        "tags": [
          "WA Base Times"
        ]
      },
      "post": {
        "operationId": "WaBaseTimesController_upsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertWaBaseTimeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create or update WA base time row (admin only)",
        "tags": [
          "WA Base Times"
        ]
      }
    },
    "/api/wa-base-times/{id}": {
      "delete": {
        "operationId": "WaBaseTimesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Delete WA base time row (admin only)",
        "tags": [
          "WA Base Times"
        ]
      }
    },
    "/api/ua-sport-ranks": {
      "get": {
        "operationId": "UaSportRanksController_findAll",
        "parameters": [
          {
            "name": "poolLength",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List UA sport rank rules, optionally by pool length",
        "tags": [
          "UA Sport Ranks"
        ]
      },
      "post": {
        "operationId": "UaSportRanksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUaSportRankDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create UA sport rank rule (admin only)",
        "tags": [
          "UA Sport Ranks"
        ]
      }
    },
    "/api/ua-sport-ranks/{id}": {
      "patch": {
        "operationId": "UaSportRanksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUaSportRankDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update UA sport rank rule (admin only)",
        "tags": [
          "UA Sport Ranks"
        ]
      },
      "delete": {
        "operationId": "UaSportRanksController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Delete UA sport rank rule (admin only)",
        "tags": [
          "UA Sport Ranks"
        ]
      }
    },
...
  },
  "info": {
    "title": "SwimSync API",
    "description": "REST API for SwimSync services",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_token"
      }
    },
    "schemas": {
...
      "UpsertWaBaseTimeRequestDto": {
        "type": "object",
        "properties": {
          "year": {
            "type": "number"
          },
          "gender": {
            "type": "string"
          },
          "distance": {
            "type": "number"
          },
          "style": {
            "type": "string"
          },
          "poolLength": {
            "type": "number"
          },
          "baseTimeMs": {
            "type": "number"
          }
        },
        "required": [
          "year",
          "gender",
          "distance",
          "style",
          "poolLength",
          "baseTimeMs"
        ]
      },
      "CreateUaSportRankDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateUaSportRankDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}